+++ infra/terraform/main.tf
@@ -22,15 +22,9 @@ provider "yandex" {
 module "test_registry" {
   source = "./modules/test_registry"
 }
-module "private_registry" {
-  source = "./modules/private_registry"
-}
 module "test_host" {
-  source                    = "./modules/test_host"
-  id_of_test_registry       = module.test_registry.test_registry_id
-  id_of_private_registry    = module.private_registry.private_registry_id
-  id_of_sa_test_registry    = module.test_registry.sa_test_registry_id
-  id_of_sa_private_registry = module.private_registry.sa_private_registry_id
-  name_sa_test_registry     = module.test_registry.name_sa_test_registry
-  name_sa_private_registry  = module.private_registry.name_sa_private_registry
+  source                 = "./modules/test_host"
+  id_of_test_registry    = module.test_registry.test_registry_id
+  id_of_sa_test_registry = module.test_registry.sa_test_registry_id
+  name_sa_test_registry  = module.test_registry.name_sa_test_registry
 }