+++ infra/terraform/yc_delete_profile.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-yc config profile activate default
-
-yc config profile delete sa-private-registry
-
-yc config profile delete sa-test-registry